move inline keyboard service methods to main bot class due to refactoring, add pictires for waste markups
